Pre-populates database with local cities list
