+++ app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/local/database/WeatherForecastDatabase.kt
@@ -47,6 +47,7 @@ abstract class WeatherForecastDatabase : RoomDatabase() {
                 WeatherForecastDatabase::class.java,
                 DB_NAME
             )
+                .createFromAsset("citiesListDatas.db")
                 .addMigrations(nToNPlusOneMigration)
                 .build()
