+++ app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/remote/model/Clouds.kt
@@ -1,0 +1,7 @@
+package ovh.geoffrey_druelle.weatherforecast.data.remote.model
+
+import com.google.gson.annotations.SerializedName
+
+data class Clouds(
+    @SerializedName("all") val all: Int
+)
