+++ app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/remote/model/Coord.kt
@@ -1,0 +1,8 @@
+package ovh.geoffrey_druelle.weatherforecast.data.remote.model
+
+import  com.google.gson.annotations.SerializedName
+
+data class Coord(
+    @SerializedName("lat") val lat: Double,
+    @SerializedName("lon") val lon: Double
+)
