+++ app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/remote/model/ListItem.kt
@@ -1,0 +1,15 @@
+package ovh.geoffrey_druelle.weatherforecast.data.remote.model
+
+import com.google.gson.annotations.SerializedName
+
+data class ListItem(
+    @SerializedName("clouds") val clouds: Clouds,
+    @SerializedName("dt") val dt: Long,
+    @SerializedName("dt_txt") val dt_txt: String,
+    @SerializedName("main") val main: Main,
+    @SerializedName("rain") val rain: Rain?,
+    @SerializedName("snow") val snow: Snow?,
+    @SerializedName("sys") val sys: Sys,
+    @SerializedName("weather") val weather: List<Weather>,
+    @SerializedName("wind") val wind: Wind
+)
