+++ app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/remote/model/Snow.kt
@@ -1,0 +1,7 @@
+package ovh.geoffrey_druelle.weatherforecast.data.remote.model
+
+import  com.google.gson.annotations.SerializedName
+
+data class Snow(
+    @SerializedName("3h") val volumeSnowLastThreeHours: Double
+)
