--- app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/remote/model/cities/CitiesCoord.kt
@@ -1,8 +1,0 @@
-package ovh.geoffrey_druelle.weatherforecast.data.remote.model.cities
-
-import com.google.gson.annotations.SerializedName
-
-data class CitiesCoord(
-    @SerializedName("lat") val lat: Double,
-    @SerializedName("lon") val lon: Double
-)
