--- app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/remote/model/openweathermap/Forecast.kt
@@ -1,11 +1,0 @@
-package ovh.geoffrey_druelle.weatherforecast.data.remote.model.openweathermap
-
-import com.google.gson.annotations.SerializedName
-
-data class Forecast(
-    @SerializedName("city") val city: City,
-    @SerializedName("cnt") val cnt: Int,
-    @SerializedName("cod") val cod: String,
-    @SerializedName("list") val list: List<ListItem>,
-    @SerializedName("message") val message: Int
-)
