--- app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/remote/model/openweathermap/Weather.kt
@@ -1,10 +1,0 @@
-package ovh.geoffrey_druelle.weatherforecast.data.remote.model.openweathermap
-
-import com.google.gson.annotations.SerializedName
-
-data class Weather(
-    @SerializedName("description") val description: String,
-    @SerializedName("icon") val icon: String,
-    @SerializedName("id") val id: Int,
-    @SerializedName("main") val main: String
-)
