--- app/src/main/java/ovh/geoffrey_druelle/weatherforecast/data/remote/model/openweathermap/Wind.kt
@@ -1,8 +1,0 @@
-package ovh.geoffrey_druelle.weatherforecast.data.remote.model.openweathermap
-
-import com.google.gson.annotations.SerializedName
-
-data class Wind(
-    @SerializedName("deg") val deg: Int,
-    @SerializedName("speed") val speed: Double
-)
