+++ app/src/main/java/ovh/geoffrey_druelle/weatherforecast/ui/splash/SplashScreenViewModel.kt
@@ -2,22 +2,19 @@ package ovh.geoffrey_druelle.weatherforecast.ui.splash
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.google.gson.GsonBuilder
-import com.google.gson.JsonArray
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.runBlocking
 import ovh.geoffrey_druelle.weatherforecast.R
 import ovh.geoffrey_druelle.weatherforecast.WeatherForecastApplication.Companion.appContext
 import ovh.geoffrey_druelle.weatherforecast.WeatherForecastApplication.Companion.instance
 import ovh.geoffrey_druelle.weatherforecast.core.BaseViewModel
-import ovh.geoffrey_druelle.weatherforecast.data.remote.api.CitiesListApi
 import ovh.geoffrey_druelle.weatherforecast.data.remote.api.OpenWeatherMapApi
-import ovh.geoffrey_druelle.weatherforecast.data.remote.model.cities.CitiesListItem
-import ovh.geoffrey_druelle.weatherforecast.data.remote.model.openweathermap.Forecast
-import ovh.geoffrey_druelle.weatherforecast.data.repository.CitiesListItemRepository
+import ovh.geoffrey_druelle.weatherforecast.data.remote.model.Forecast
 import ovh.geoffrey_druelle.weatherforecast.data.repository.CityRepository
 import ovh.geoffrey_druelle.weatherforecast.data.repository.ForecastRepository
 import ovh.geoffrey_druelle.weatherforecast.utils.cleanForecastDatabase
-import ovh.geoffrey_druelle.weatherforecast.utils.createCitiesListObject
 import ovh.geoffrey_druelle.weatherforecast.utils.createCityObject
 import ovh.geoffrey_druelle.weatherforecast.utils.createForecastObject
 import ovh.geoffrey_druelle.weatherforecast.utils.helper.ConnectivityHelper.isConnectedToNetwork
@@ -27,9 +24,9 @@ import retrofit2.Response
 import timber.log.Timber
 import kotlin.coroutines.CoroutineContext
 
+
 class SplashScreenViewModel(
-    private val owmApi: OpenWeatherMapApi,
-    private val cityApi: CitiesListApi
+    private val owmApi: OpenWeatherMapApi
 ) : BaseViewModel(), CoroutineScope {
 
     override val coroutineContext: CoroutineContext
@@ -39,7 +36,6 @@ class SplashScreenViewModel(
 
     private var forecastRepository = ForecastRepository(instance)
     private var cityRepository = CityRepository(instance)
-    private var citiesListRepository = CitiesListItemRepository(instance)
 
     private val _isConnected = MutableLiveData<Boolean>()
     val isConnected: LiveData<Boolean>
@@ -73,21 +69,7 @@ class SplashScreenViewModel(
             false
         }
 
-        checkCitiesListDatas(bool)
-    }
-
-    private fun checkCitiesListDatas(bool: Boolean) {
-        val citiesListCount = runBlocking {
-            citiesListRepository.getLength()
-        }
-
-        when {
-            bool && citiesListCount != 0 -> checkForecastDatas(bool)
-            bool && citiesListCount == 0 -> {
-                launchRequestForCitiesList(bool)
-            }
-            else -> _noDataNoConnection.postValue(true)
-        }
+        checkForecastDatas(bool)
     }
 
     private fun checkForecastDatas(bool: Boolean) {
@@ -99,58 +81,6 @@ class SplashScreenViewModel(
             bool -> launchRequestForForecastDatas()
             forecastCount != 0 -> _navToHome.postValue(true)
             else -> _noDataNoConnection.postValue(true)
-        }
-    }
-
-    private fun launchRequestForCitiesList(bool: Boolean) {
-        val call: Call<JsonArray> = cityApi.readJson()
-        call.enqueue(object : Callback<JsonArray> {
-            override fun onFailure(call: Call<JsonArray>, t: Throwable) {
-                Timber.d(String.format("launchRequestForCitiesList : Failure on call - %s", t))
-                Timber.d(
-                    String.format(
-                        "launchRequestForCitiesList : Failure on call - %s",
-                        t.message
-                    )
-                )
-            }
-
-            override fun onResponse(call: Call<JsonArray>, response: Response<JsonArray>) {
-                val jsonArray = response.body() as JsonArray
-                if (response.isSuccessful) {
-                    getCitiesListFromJsonArray(jsonArray, bool)
-                } else {
-                    Timber.d(String.format("launchRequestForCitiesList: got response but not successful"))
-                }
-            }
-        })
-    }
-
-    private fun getCitiesListFromJsonArray(
-        jsonArray: JsonArray,
-        bool: Boolean
-    ) {
-        runBlocking {
-            withContext(Dispatchers.IO) {
-                val gsonBuilder = GsonBuilder().serializeNulls()
-                val gson = gsonBuilder.create()
-                val citiesListItem: List<CitiesListItem> =
-                    gson.fromJson(jsonArray, Array<CitiesListItem>::class.java).toList()
-                populateCitiesListDatabase(citiesListItem, bool)
-            }
-        }
-    }
-
-    private fun populateCitiesListDatabase(
-        citiesListItem: List<CitiesListItem>,
-        bool: Boolean
-    ) {
-        runBlocking {
-            for (i in citiesListItem.indices) {
-                citiesListRepository.insert(createCitiesListObject(citiesListItem[i]))
-            }
-            citiesListRepository.deleteContinents()
-            checkForecastDatas(bool)
         }
     }
