+++ app/src/main/java/ovh/geoffrey_druelle/weatherforecast/utils/Constants.kt
@@ -10,4 +10,3 @@ const val WRITE_TIMEOUT = 15L
 const val READ_TIMEOUT = 15L
 // Base of the url of used API
 const val BASE_URL = "http://api.openweathermap.org/data/2.5/"
-const val CITIES_URL = "https://geoffrey-druelle.ovh/res/"
